+++ CHANGELOG.md
@@ -7,6 +7,14 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
 
+### Changed
+- Fixed http-json example outputs
+
+## [0.3.0] - 2021-01-20
+
+### BREAKING CHANGE
+- Changed from --path using gojsonq to --query using gojq
+
 ## [0.2.0] - 2021-01-13
 
 ### Changed